refactor: 회원 서비스 수정(#177)
