--- src/main/java/com/leeforgiveness/memberservice/auth/dto/MemberHandleRequestDto.java
@@ -1,17 +1,0 @@
-package com.leeforgiveness.memberservice.auth.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class MemberHandleRequestDto {
-
-	private String uuid;
-}
