--- src/main/java/com/leeforgiveness/memberservice/auth/dto/MemberHandleResponseDto.java
@@ -1,10 +1,0 @@
-package com.leeforgiveness.memberservice.auth.dto;
-
-import com.leeforgiveness.memberservice.auth.vo.MemberHandleResponseVo;
-
-public class MemberHandleResponseDto {
-
-	private String handle;
-
-//	public static MemberHandleResponseVo dtoToVo
-}
