+++ src/main/java/com/leeforgiveness/memberservice/auth/dto/SellerMemberDetailResponseDto.java
@@ -17,21 +17,13 @@ import lombok.Setter;
 public class SellerMemberDetailResponseDto {
 
     private String name;
-    private String handle;
-    private List<String> watchList;
     private String profileImage;
-    private List<Map<String, Object>> careerInfo;
-    private List<Map<String, Object>> qualificationInfo;
     private boolean isSubscribed;
 
     public static SellerMemberDetailResponseVo dtoToVo(
         SellerMemberDetailResponseDto sellerMemberDetailResponseDto) {
         return new SellerMemberDetailResponseVo(
             sellerMemberDetailResponseDto.getName(),
-            sellerMemberDetailResponseDto.getCareerInfo(),
-            sellerMemberDetailResponseDto.getQualificationInfo(),
-            sellerMemberDetailResponseDto.getHandle(),
-            sellerMemberDetailResponseDto.getWatchList(),
             sellerMemberDetailResponseDto.getProfileImage(),
             sellerMemberDetailResponseDto.isSubscribed());
     }
