--- src/main/java/com/leeforgiveness/memberservice/auth/infrastructure/InterestCategoryRepository.java
@@ -1,12 +1,0 @@
-package com.leeforgiveness.memberservice.auth.infrastructure;
-
-import com.leeforgiveness.memberservice.auth.domain.InterestCategory;
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface InterestCategoryRepository extends JpaRepository<InterestCategory, Long> {
-
-    List<InterestCategory> findByUuid(String uuid);
-}
