+++ src/main/java/com/leeforgiveness/memberservice/auth/infrastructure/MemberRepository.java
@@ -13,8 +13,6 @@ public interface MemberRepository extends JpaRepository<Member, Long> {
 
     Optional<Member> findByUuid(String uuid);
 
-    Optional<Member> findByHandle(String handle);
-
     List<Member> findByUuidIn(List<String> uuids);
 
     Optional<Member> findByPhoneNum(String phoneNum);
