--- src/main/java/com/leeforgiveness/memberservice/auth/vo/MemberCareerAddRequestVo.java
@@ -1,11 +1,0 @@
-package com.leeforgiveness.memberservice.auth.vo;
-
-import lombok.Getter;
-
-@Getter
-public class MemberCareerAddRequestVo {
-
-	private String job;
-	private int year;
-	private int month;
-}
