--- src/main/java/com/leeforgiveness/memberservice/auth/vo/MemberCareerDeleteRequestVo.java
@@ -1,10 +1,0 @@
-package com.leeforgiveness.memberservice.auth.vo;
-
-import lombok.Getter;
-
-@Getter
-public class MemberCareerDeleteRequestVo {
-
-	private String job;
-
-}
