--- src/main/java/com/leeforgiveness/memberservice/auth/vo/MemberHandleRequestVo.java
@@ -1,6 +1,0 @@
-package com.leeforgiveness.memberservice.auth.vo;
-
-public class MemberHandleRequestVo {
-
-	private String uuid;
-}
