--- src/main/java/com/leeforgiveness/memberservice/auth/vo/MemberHandleResponseVo.java
@@ -1,6 +1,0 @@
-package com.leeforgiveness.memberservice.auth.vo;
-
-public class MemberHandleResponseVo {
-
-	private String handle;
-}
