--- src/main/java/com/leeforgiveness/memberservice/auth/vo/MemberQualificationDeleteRequestVo.java
@@ -1,9 +1,0 @@
-package com.leeforgiveness.memberservice.auth.vo;
-
-import lombok.Getter;
-
-@Getter
-public class MemberQualificationDeleteRequestVo {
-
-	private String name;
-}
