+++ src/main/java/com/leeforgiveness/memberservice/auth/vo/MemberUpdateRequestVo.java
@@ -7,6 +7,5 @@ public class MemberUpdateRequestVo {
 
 	public String name;
 	public String phoneNum;
-	public String handle;
 	public String profileImage;
 }
