+++ src/main/java/com/leeforgiveness/memberservice/auth/vo/SellerMemberDetailRequestVo.java
@@ -10,5 +10,4 @@ import lombok.Setter;
 public class SellerMemberDetailRequestVo {
 
 	String uuid;
-	String handle;
 }
